+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.2.0 - 2026-04-22
+
+- Log warning if an empty config is generated
+- Log warning if a client connects when no config is available
+- Add more detail to debug logs
+- If no valid containers are found, the empty config is now sent to Centauri.
+
 ## 1.1.0 - 2026-04-04
 
 - Dependency updates